+++ src/main/java/lotto/model/Interface/WinResultStorage.java
@@ -1,0 +1,8 @@
+package lotto.model.Interface;
+
+public interface WinResultStorage {
+
+  int[] getWinResult();
+
+  double getProfitRate();
+}
