--- src/main/java/lotto/model/WinResultStorage.java
@@ -1,8 +1,0 @@
-package lotto.model;
-
-interface WinResultStorage {
-
-  int[] getWinResult();
-
-  double getProfitRate();
-}
